Refactor import paths in ParaglideMetaTags.svelte
The path has been updated for importing 'translatePath' in ParaglideMetaTags.svelte due to a change in the name of the file where it's situated. 'index.ts' file has been renamed to 'utils.ts', hence modifying the import path to ensure the correct import of functions and avoid any potential breaks.